+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ### Added
+- [Minor] ISDU Access for CPX-AP-4IOL
+- [Minor] Added CPX-AP-A modules
 - [Minor] Implemented CPX-E-1CI module
 - [Minor] New functions for CPX-E-4AI and 4AO
 - [Minor] Added utils/helpers with ceil function
@@ -23,6 +25,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [Minor] CpxE8Do: Added access-by-index-operator to channels
 - [Minor] CpxE: Added CLI for writing and reading values.
 ### Changed
+- [Minor] Moved typecode ids to cpx_*_definitions.py
+- [Minor] e1ci dicts are now dataclasses
 - [Minor] Modbus Commands are now marked as constants
 - [Minor] Moved modules to individual files
 - [Minor] Moved unittests to subfolder